refactor: remove outdated DaisyUI preview components and streamline codebase
- Delete unused layout, page, and component files related to DaisyUI preview.
- Update the usePreview hook to reflect changes in the preview URL structure.
- Introduce new components and utilities for improved theme generation and customization.
- Enhance the overall structure and maintainability of the codebase by removing deprecated files.
