--- src/app/daisy-preview/layout.tsx
@@ -1,7 +1,0 @@
-export default function DaisyPreviewLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode
-}>) {
-  return <div className="daisy-preview-layout">{children}</div>
-}
