+++ src/app/wizard/components/color-palette.tsx
