--- src/app/wizard/utils/generate-tailwind-v4-css.ts
@@ -1,29 +1,0 @@
-import { generateTailwindV4CSS as generateTheme } from '../themes/tailwind-theme'
-import { type BorderRadius, type ButtonRadius, type ButtonStyle } from '../types'
-
-export interface TailwindConfig {
-  primaryColor: string
-  secondaryColor: string
-  accentColor: string
-  headingFont: string
-  bodyFont: string
-  monoFont: string
-  borderRadius?: BorderRadius
-  spacing?: string
-  buttonStyle?: ButtonStyle
-  buttonRadius?: ButtonRadius
-  primaryShades?: Record<string, string>
-  secondaryShades?: Record<string, string>
-  accentShades?: Record<string, string>
-  primaryShade?: string
-  secondaryShade?: string
-  accentShade?: string
-}
-
-/**
- * Generate complete Tailwind CSS v4 configuration
- * This is the main entry point for the CSS generation
- */
-export function generateTailwindV4CSS(config: TailwindConfig): string {
-  return generateTheme(config)
-}
