--- src/app/wizard/utils/normalize-hue.ts
@@ -1,6 +1,0 @@
-/**
- * Normalize hue to the range [0, 360)
- */
-export function normalizeHue(hue: number): number {
-  return ((hue % 360) + 360) % 360
-}
